[Decomp][Item] Cohesion split: extract vendor/bank handlers to ItemHandlerVendor.cpp
